requirements.txt
streamlit
crewai[google-genai]>=0.1.0
google-generativeai>=0.8.0
crewai-tools
python-dotenv
litellm
